requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
Rx==3.2.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
